+++ webpack.config.js
@@ -1,5 +1,9 @@
 var Encore = require( '@symfony/webpack-encore' );
 
+if ( ! Encore.isRuntimeEnvironmentConfigured() ) {
+    Encore.configureRuntimeEnvironment( process.env.NODE_ENV || 'dev' );
+}
+
 /**
  *  AdminPanel Default Theme
  */